backend/pyproject.toml
[project]
name = "mini-knowledge-graph-backend"
version = "1.0.0"
description = "FastAPI backend for Mini Knowledge Graph"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "sqlalchemy>=2.0.35",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    "pdfplumber>=0.11.4",
    "langchain-groq>=0.2.0",
    "langchain-core>=0.3.0",
    "alembic>=1.13.2",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.28.0",
    "redis>=5.0.0",
    "arq>=0.25.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.11.0",
    "bcrypt>=5.0.0",
]
